+++ crates/runtime/src/dialogue.rs
@@ -1,11 +1,6 @@
-pub use self::{dialogue_option::*, handlers::*, line::*, variable_storage::*};
+use crate::prelude::*;
 use std::fmt::Debug;
 use yarn_slinger_core::prelude::*;
-
-mod dialogue_option;
-mod handlers;
-mod line;
-mod variable_storage;
 
 /// Co-ordinates the execution of Yarn programs.
 #[non_exhaustive]
