+++ crates/runtime/src/lib.rs
@@ -1,5 +1,9 @@
 mod dialogue;
+mod dialogue_option;
+mod handlers;
+mod line;
+mod variable_storage;
 
 pub mod prelude {
-    pub use crate::dialogue::*;
+    pub use crate::{dialogue::*, dialogue_option::*, handlers::*, line::*, variable_storage::*};
 }
